requirements.txt
# ComfyUI Model Preset Pilot Requirements
# Dependencies for the Model Preset Pilot module

torch>=1.9.0
pillow>=8.3.0
numpy>=1.21.0
